src/BusinessProcessAgent.Agent/requirements.txt
agent-framework-azure-ai==1.0.0rc3
agent-framework-core==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16
azure-identity>=1.19.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
